♻️(front) configure dashjs to start at a lowest bitrate
The bitrate expected in the updateSettings method is in kbps and not in
bps, changing this value allow us to start a video in 480p. We also decided
to isolate dash creation in a module. Existing test with real dash
instance have been removed because dash throw an error it expects a real
video to parse.
